Add files via upload
The data used for Figure 10 and Figure 11 in the main text is provided in this Excel file. The sheet titled “Figure 10 (correlation)” shows the data used to plot Figure 10. For Figure 11, the source of the data is provided in “Figure 11 (Data source)”, the data used to fit correlations in Figure 11a is summarized in “Figure 11 (correlation)”, and the data used to plot Figure 11b and 11c are provided in “Figure 11 (comparison)”.